Update Gilded RP Rules.md
Criminals may not wear all black clothing with 0 identifiable features while committing crime. They must wear at least 1-3 identifiable features, whether this be a buckle, bandana, clothing item, gun, or horse and saddle. Failure to do so is abuse of the voice ID rule and is a method of Powergaming to always win a situation so nobody may catch, or pursue you in retaliation for the crime.

Signed-off-by: DJRLincs <90622354+DJRLincs@users.noreply.github.com>