Handcuffs and Wagon Update
Added clause for wagon robbery rules to apply to horse saddlebags to stop confusion over that rule, 

Added handcuff rules to require the victim to have hands raised, or have their back to the aggressor. Instant cuffing noted to be likened to low-tier initiation and instant roping in this regard.

Few formatting and grammar changes that have been pissing me off forever.

Signed-off-by: Haldor-Hammers <adfdiabolus@gmail.com>